Update README.md
Extended information on the `Execute` function. Not looking into the internal code at all I was not sure based on the README what happen to the data passed to the `Dispatch` and `Invoke` functions.

Due to this I just thought id add a sentence denoting what the parameters/arguments are for and where you can access them in case it was not obvious.

Could also be moved into the ParallelWorker part of the docs.